Iss2214 - Update and retrieve run log line count in Directory and CouchDB RAS store (#359)
* feat: First commit, retrieve log so far from CouchDB during test to count run log lines


* feat: Update the run log line count when writing to it and retrieve the run log line count from CouchDB, to add to test structure for each method


* feat: Change type of runLogStart and runLogEnd to long to protect against edge case of run log containing more lines than the maximum int value


---------

+++ modules/framework/galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/common/RunResultUtility.java
@@ -85,8 +85,8 @@ public class RunResultUtility {
       	      String result = method.getResult();
       	      Instant startTime = method.getStartTime();
       	      Instant endTime = method.getEndTime();
-      	      int runLogStart = method.getRunLogStart();
-      	      int runLogEnd = method.getRunLogEnd();
+      	      long runLogStart = method.getRunLogStart();
+      	      long runLogEnd = method.getRunLogEnd();
       	      List<RasTestMethod> befores = convertMethods(method.getBefores());
       	      List<RasTestMethod> afters = convertMethods(method.getAfters());
