+++ modules/framework/galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/GenericMethodWrapper.java
@@ -62,7 +62,7 @@ public class GenericMethodWrapper {
      */
     public void invoke(@NotNull ITestRunManagers managers, Object testClassObject, GenericMethodWrapper testMethod, TestClassWrapper testClassWrapper) throws TestRunException {
 
-        int runLogStart = testClassWrapper.getRunLogLineCount();
+        long runLogStart = testClassWrapper.getRunLogLineCount();
         
         try {
             // Associate the wrapped method with a test method if a test method has been passed in
@@ -135,7 +135,7 @@ public class GenericMethodWrapper {
             throw new TestRunException("There was a problem with the framework, please check stacktrace", e);
         }
 
-        int runLogEnd = testClassWrapper.getRunLogLineCount();
+        long runLogEnd = testClassWrapper.getRunLogLineCount();
 
         // Compare the run log start and run log end to see if this method produced any output.
         // If it did then set the runLogStart and runLogEnd in the test structure.
@@ -166,11 +166,11 @@ public class GenericMethodWrapper {
         }
     }
 
-    public void setRunLogStart(int runLogStart) {
+    public void setRunLogStart(long runLogStart) {
         this.testStructureMethod.setRunLogStart(runLogStart);
     }
 
-    public void setRunLogEnd(int runLogEnd) {
+    public void setRunLogEnd(long runLogEnd) {
         this.testStructureMethod.setRunLogEnd(runLogEnd);
     }
