+++ modules/framework/galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/spi/teststructure/TestGherkinMethod.java
@@ -16,8 +16,8 @@ public class TestGherkinMethod {
 
     private String           exception;
 
-    private int              runLogStart;
-    private int              runLogEnd;
+    private long              runLogStart;
+    private long              runLogEnd;
 
     private Instant          startTime;
     private Instant          endTime;
@@ -68,19 +68,19 @@ public class TestGherkinMethod {
         this.exception = exception;
     }
 
-    public int getRunLogStart() {
+    public long getRunLogStart() {
         return runLogStart;
     }
 
-    public void setRunLogStart(int runLogStart) {
+    public void setRunLogStart(long runLogStart) {
         this.runLogStart = runLogStart;
     }
 
-    public int getRunLogEnd() {
+    public long getRunLogEnd() {
         return runLogEnd;
     }
 
-    public void setRunLogEnd(int runLogEnd) {
+    public void setRunLogEnd(long runLogEnd) {
         this.runLogEnd = runLogEnd;
     }
